requirements.txt
# Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
# gunicorn==23.0.0   # ← opcional; elimínalo si usarás uvicorn directo

# LangChain y ecosistema
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
langchain-openai==0.2.0
openai==1.40.0

# Embeddings y vector DB
chromadb==0.5.5
tiktoken==0.7.0

# PDF y documentos
pypdf==4.2.0

# Estáticos / utilidades
aiofiles==23.2.1
python-dotenv==1.0.1
tqdm==4.66.5
requests==2.32.3


